Update app.js
If the application runs somewhere else than radio.canoo.com or the webpage is opened via the IP, the current implementation fails to provide a usable player. The change should solve it, but limits CanooRadio to have the client UI and the streaming server running on the same machine. 
Maybe we need to come up with a better attempt e.g. for development time, etc.